+++ core/Libraries.php
@@ -992,7 +992,7 @@ class Libraries {
 		$suffix = $options['namespaces'] ? '' : $config['suffix'];
 		$suffix = ($options['suffix'] === null) ? $suffix : $options['suffix'];
 
-		$dFlags = GLOB_ONLYDIR & GLOB_BRACE;
+		$dFlags = GLOB_ONLYDIR | GLOB_BRACE;
 		$libs = (array) glob($path . $suffix, $options['namespaces'] ? $dFlags : GLOB_BRACE);
 
 		if ($options['recursive']) {